Default Eigen in CMake, add tests.
Default the Eigen library in CMake, as it's the fastest for most
contemporary CPUs and configurations, and the easiest to build.

We can optionally use BLAS by adding the USE_BLAS define, and
will try to locate BLAS/OpenBLAS if so. This is useful for
binaries for distribution such as our releases or distros.

Split all tests to cover both Eigen and BLAS.